app*:: folders/files structure refactoring (components---pages---layouts)
