--- src/components/DetailPage/index.ts
@@ -1,3 +1,0 @@
-import { DetailPage } from "./DetailPage";
-
-export default DetailPage;
