--- src/components/HomePage/index.ts
@@ -1,3 +1,0 @@
-import { HomePage } from "./HomePage";
-
-export default HomePage;
