--- src/components/SettingsPage/index.ts
@@ -1,3 +1,0 @@
-import { SettingsPage } from "./SettingsPage";
-
-export default SettingsPage;
