+++ src/pages/DetailPage/index.ts
@@ -1,0 +1,1 @@
+export * from "./DetailPage";
