+++ src/pages/HomePage/HomePage.tsx
@@ -1,7 +1,7 @@
 import React, {FC, useEffect} from "react";
 import {DESCRIPTION_SYMBOLS_QUANTITY, ROUTES_PATH} from "../../constants";
 import {IPost} from "../../types/post";
-import ProgressBar from "../ProgressBar";
+import {ProgressBar} from "../../components";
 import {Link} from "react-router-dom";
 import {renderTags} from "../../helpers/renderFuncs";
 import InfiniteScroll from "react-infinite-scroll-component";
@@ -93,4 +93,4 @@ export const HomePage: FC = () => {
             {isLoading ? (<ProgressBar/>) : ''}
         </>
     )
-}
+};
