+++ src/pages/index.ts
@@ -1,0 +1,4 @@
+export * from './DetailPage';
+export * from './HomePage';
+export * from './LoginPage';
+export * from './SettingsPage';
